:recycle: refactor: slugify as client utils
